Only unify top-level package.json and let npm resolve any conflicts found.
Instead of trying to use semver to unify different package versions, just
use the `and` operator (in semver, a space) to combine the versions and
let npm figure it out.  Similarly, don't try to unify *all* the
package.json files; just do the top-level ones and let `npm dedupe` handle
the recursion.

Change-Id: I968afa24edd00c5574317c974749cd54a792ac3f